+++ src/app/(pages)/(search)/(home)/layout.tsx
@@ -1,4 +1,4 @@
-import { Footer } from '@/app/components/parts/Footer/Footer';
+import { Footer } from '@/app/_components/parts/Footer/Footer';
 
 export default function FooterLayout({ children }: { children: React.ReactNode }) {
   return (
