+++ src/app/(pages)/fav/layout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from '@/app/components/parts/Footer/Footer';
-import { PageHead } from '@/app/components/parts/PageHead';
+import { Footer } from '@/app/_components/parts/Footer/Footer';
+import { PageHead } from '@/app/_components/parts/PageHead';
 
 export default function FavLayout({ children }: { children: React.ReactNode }) {
   return (
