+++ src/app/(pages)/list/layout.tsx
@@ -1,5 +1,5 @@
-import { Footer } from '@/app/components/parts/Footer/Footer';
-import { PageHead } from '@/app/components/parts/PageHead';
+import { Footer } from '@/app/_components/parts/Footer/Footer';
+import { PageHead } from '@/app/_components/parts/PageHead';
 
 export default function ListLayout({ children }: { children: React.ReactNode }) {
   return (
