+++ src/app/_components/features/recipe/RecipeList/RecipeListVertical.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 
-import { dummyRecipeList } from '@/app/data/dummyRecipe';
-import { Recipe } from '@/app/types/Recipe';
+import { dummyRecipeList } from '@/app/_data/dummyRecipe';
+import { Recipe } from '@/app/_types/Recipe';
 
 import { RecipeListItem } from './RecipeListItem';
 
@@ -9,12 +9,12 @@ type Props = {
   recipeList: Recipe[];
 };
 
-export const RecipeListHorizontal: FC<Props> = ({ recipeList }) => {
+export const RecipeListVertical: FC<Props> = ({ recipeList }) => {
   const recipe = recipeList.length > 0 ? recipeList : dummyRecipeList;
   return (
-    <ul className="hidden-scrollbar flex gap-x-4 overflow-x-scroll">
+    <ul className="grid grid-cols-2  gap-x-3 gap-y-4">
       {recipe.map((recipe) => (
-        <li key={recipe.id} className="w-40 flex-shrink-0">
+        <li key={recipe.id} className="">
           <RecipeListItem recipe={recipe} />
         </li>
       ))}
