+++ src/app/_components/parts/Footer/Icons/propsType.ts
