+++ src/app/_components/parts/Footer/index.ts
