+++ src/app/_components/parts/PageHead/index.ts
