+++ src/app/_data/dummyRecipe.ts
@@ -1,4 +1,4 @@
-import { Recipe } from '../types/Recipe';
+import { Recipe } from '../_types/Recipe';
 
 // 日本語でダミーレシピを作成
 export const dummyRecipeList: Recipe[] = [
